requirements.txt
streamlit
streamlit_extras
python-dotenv
PyPDF2
Pillow
google.generativeai